\name{GINA.sample}
\alias{GINA.sample}
\title{cranberry image}
\description{

Sample image from the cranberry genomics program. The image displays 25 cranberry fruits from a single genotyped captured by a camera in a white background with black reference circles for standarization purposes.

}

\usage{
  data(GINA.sample)
}

\format{
 An image with 3 layers for green, red and blue. 
}

\source{
  Cranberry Genomics Program, Madison, WI, USA.
}

\references{
Diaz-Garcia L, Covarrubias-Pazaran G, Schlautman B, Salazar W, Zalapa J. (2015) GINA: A flexible high throughput phenotyping tool. http://horticulture.wisc.edu/cggl/ZalapaLab/People.html. 2015.

Ben Hui Liu. 1998. Statistical Genomics. CRC Press LLC. ISBN 0-8493-3166-8.
}
\examples{
####=========================================####
#### For CRAN time limitations most lines in the 
#### examples are silenced with one '#' mark, 
#### remove them and run the examples
####=========================================####
data(GINA.sample)
#display(GINA.sample)
#str(GINA.sample)
}
\keyword{datasets}
